// src/lib/openapi/region/models/LoadBalancerListenerV2.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Region Service API
 * Cloud region discovery and routing service.  This is service not intended for direct access by end users.  Region related functionality is typically exposed by higher level services e.g. Compute and Kubernetes, that provide subsets of resources that are compatible with that service.
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { LoadBalancerListenerProtocolV2 } from './LoadBalancerListenerProtocolV2';
import {
    LoadBalancerListenerProtocolV2FromJSON,
    LoadBalancerListenerProtocolV2FromJSONTyped,
    LoadBalancerListenerProtocolV2ToJSON,
} from './LoadBalancerListenerProtocolV2';
import type { LoadBalancerPoolV2 } from './LoadBalancerPoolV2';
import {
    LoadBalancerPoolV2FromJSON,
    LoadBalancerPoolV2FromJSONTyped,
    LoadBalancerPoolV2ToJSON,
} from './LoadBalancerPoolV2';

/**
 * A load balancer listener.
 * @export
 * @interface LoadBalancerListenerV2
 */
export interface LoadBalancerListenerV2 {
    /**
     * A load balancer listener name. Must start with a lower-case letter and otherwise be a valid DNS label.
     * @type {string}
     * @memberof LoadBalancerListenerV2
     */
    name: string;
    /**
     * 
     * @type {LoadBalancerListenerProtocolV2}
     * @memberof LoadBalancerListenerV2
     */
    protocol: LoadBalancerListenerProtocolV2;
    /**
     * The listener port.
     * @type {number}
     * @memberof LoadBalancerListenerV2
     */
    port: number;
    /**
     * A list of source IPv4 CIDRs allowed to access the listener.
     * @type {Array<string>}
     * @memberof LoadBalancerListenerV2
     */
    allowedCidrs?: Array<string>;
    /**
     * The TCP idle timeout in seconds.
     * @type {number}
     * @memberof LoadBalancerListenerV2
     */
    idleTimeoutSeconds?: number;
    /**
     * 
     * @type {LoadBalancerPoolV2}
     * @memberof LoadBalancerListenerV2
     */
    pool: LoadBalancerPoolV2;
}

/**
 * Check if a given object implements the LoadBalancerListenerV2 interface.
 */
export function instanceOfLoadBalancerListenerV2(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "protocol" in value;
    isInstance = isInstance && "port" in value;
    isInstance = isInstance && "pool" in value;

    return isInstance;
}

export function LoadBalancerListenerV2FromJSON(json: any): LoadBalancerListenerV2 {
    return LoadBalancerListenerV2FromJSONTyped(json, false);
}

export function LoadBalancerListenerV2FromJSONTyped(json: any, ignoreDiscriminator: boolean): LoadBalancerListenerV2 {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'protocol': LoadBalancerListenerProtocolV2FromJSON(json['protocol']),
        'port': json['port'],
        'allowedCidrs': !exists(json, 'allowedCidrs') ? undefined : json['allowedCidrs'],
        'idleTimeoutSeconds': !exists(json, 'idleTimeoutSeconds') ? undefined : json['idleTimeoutSeconds'],
        'pool': LoadBalancerPoolV2FromJSON(json['pool']),
    };
}

export function LoadBalancerListenerV2ToJSON(value?: LoadBalancerListenerV2 | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'protocol': LoadBalancerListenerProtocolV2ToJSON(value.protocol),
        'port': value.port,
        'allowedCidrs': value.allowedCidrs,
        'idleTimeoutSeconds': value.idleTimeoutSeconds,
        'pool': LoadBalancerPoolV2ToJSON(value.pool),
    };
}